+++ src/fiabilipym/__init__.py
@@ -1,9 +1,17 @@
 """Reliability system design and computation."""
 
 from .component import Component
+from .distribution import Weibull, weibull_eta_from_lambda
 from .voter import Voter
 from .system import System
 from .markov import Markovprocess
 
 __version__ = "2.8.1"
-__all__ = ["System", "Component", "Voter", "Markovprocess"]
+__all__ = [
+    "System",
+    "Component",
+    "Voter",
+    "Markovprocess",
+    "Weibull",
+    "weibull_eta_from_lambda",
+]